Rename folder from reolink to reolink_dev
